requirements.txt
Flask==3.0.2
requests==2.31.0
fpdf2==2.7.7